backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
neo4j==5.19.0
python-dotenv==1.0.1
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
